refactory: fefactory build gradle version to querydsl

+++ src/main/java/single/project/stakeapi/application/model/transfer/response/StakeMetaDataResponse.java
@@ -3,6 +3,7 @@ package single.project.stakeapi.application.model.transfer.response;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.experimental.FieldDefaults;
+import single.project.stakeapi.application.model.entity.StakePool;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
@@ -11,11 +12,12 @@ import java.util.Map;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class StakeMetaDataResponse {
-    BigInteger stakePoolId;
+    Long stakePoolId;
     String contractName;
     String contractAddress;
     String ownerAddress;
     Map<String, String> stakeToken;
     Map<String, String> rewardToken;
     BigDecimal totalSupply;
+
 }
